# src/utils/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Course API",
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/public",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/lesson/course/{courseId}": {
      "post": {
        "summary": "create lesson",
        "tags": [
          "lesson"
        ],
        "operationId": "createlesson",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createlessonrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/createlesson"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "16"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"10-oV4hJxRVSENxc/wX8+mA4/Pe4tA\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:44:55 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      },
      "get": {
        "summary": "get all lessons by course id",
        "tags": [
          "lesson"
        ],
        "operationId": "getalllessonsbycourseid",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/getalllessonsbycourseid"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true,
                "result": [
                  {
                    "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory",
                    "lessonNumber": 1,
                    "description": "learn firs kompuktor",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T17:50:09.000Z",
                    "updatedAt": "2023-02-04T17:50:09.000Z"
                  },
                  {
                    "id": "1bb84a91-c217-4646-8ce6-a491f0c501ea",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory 2",
                    "lessonNumber": 2,
                    "description": "learn second kompuktor",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T17:50:09.000Z",
                    "updatedAt": "2023-02-04T17:50:09.000Z"
                  },
                  {
                    "id": "dafd076c-cd23-42d2-aa19-c1be0fb112f0",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory 3",
                    "lessonNumber": 3,
                    "description": "learn kompuktor of 3 gen",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T17:50:09.000Z",
                    "updatedAt": "2023-02-04T17:50:09.000Z"
                  },
                  {
                    "id": "f1c4c340-3171-487c-ae4c-425247613457",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory 4",
                    "lessonNumber": 4,
                    "description": "learn kompuktor of 4 gen",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T17:50:09.000Z",
                    "updatedAt": "2023-02-04T17:50:09.000Z"
                  },
                  {
                    "id": "15304177-d4c5-4dd5-8600-de1bc3376aa3",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory 5",
                    "lessonNumber": 5,
                    "description": "learn kompuktor of 5 gen",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T17:50:09.000Z",
                    "updatedAt": "2023-02-04T17:50:09.000Z"
                  },
                  {
                    "id": "e434a0c7-2075-4be1-bead-ccff61dc5c17",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory",
                    "lessonNumber": 6,
                    "description": "learn 6 kompuktor",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T18:39:57.000Z",
                    "updatedAt": "2023-02-04T18:39:57.000Z"
                  },
                  {
                    "id": "0aa46dcb-52ff-41e1-bf13-05dc9df044b6",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory",
                    "lessonNumber": 7,
                    "description": "learn 6 kompuktor",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T18:40:22.000Z",
                    "updatedAt": "2023-02-04T18:40:22.000Z"
                  },
                  {
                    "id": "2709083e-85e2-41ba-aab6-c714a109ea0f",
                    "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                    "title": "kompuktory",
                    "lessonNumber": 8,
                    "description": "learn 6 kompuktor",
                    "information": [
                      "information about kompuktory"
                    ],
                    "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                    "updatedBy": null,
                    "createdAt": "2023-02-04T18:44:55.000Z",
                    "updatedAt": "2023-02-04T18:44:55.000Z"
                  }
                ]
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "2952"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"b88-Ifu40TWOK2kJQGTBCmp1qjRudMM\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:45:56 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      }
    },
    "/lesson/{lessonId}/course/{courseId}": {
      "get": {
        "summary": "get lesson by Id",
        "tags": [
          "lesson"
        ],
        "operationId": "getlessonbyId",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/getlessonbyId"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true,
                "result": {
                  "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
                  "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
                  "title": "kompuktory",
                  "lessonNumber": 1,
                  "description": "learn firs kompuktor",
                  "information": [
                    "information about kompuktory"
                  ],
                  "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
                  "updatedBy": null,
                  "createdAt": "2023-02-04T17:50:09.000Z",
                  "updatedAt": "2023-02-04T17:50:09.000Z"
                }
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "389"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"185-jxq++j7e5fzHF3k4OwbAg52XFVc\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:49:32 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "update lesson",
        "tags": [
          "lesson"
        ],
        "operationId": "updatelesson",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updatelessonrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/updatelesson"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "16"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"10-oV4hJxRVSENxc/wX8+mA4/Pe4tA\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:54:21 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "summary": "delete lesson",
        "tags": [
          "lesson"
        ],
        "operationId": "deletelesson",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/deletelesson"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "16"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"10-oV4hJxRVSENxc/wX8+mA4/Pe4tA\""
              },
              "Date": {
                "type": "string",
                "default": "Sat, 04 Feb 2023 18:57:23 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "definitions": {
...
    "createlessonrequest": {
      "title": "createlessonrequest",
      "example": {
        "lessonNumber": 1,
        "title": "kompuktory",
        "description": "learn 6 kompuktor",
        "information": [
          "information about kompuktory"
        ]
      },
      "type": "object",
      "properties": {
        "lessonNumber": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "lessonNumber",
        "title",
        "description",
        "information"
      ]
    },
    "createlesson": {
      "title": "createlesson",
      "example": {
        "success": true
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      },
      "required": [
        "success"
      ]
    },
    "getalllessonsbycourseid": {
      "title": "getalllessonsbycourseid",
      "example": {
        "success": true,
        "result": [
          {
            "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory",
            "lessonNumber": 1,
            "description": "learn firs kompuktor",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T17:50:09.000Z",
            "updatedAt": "2023-02-04T17:50:09.000Z"
          },
          {
            "id": "1bb84a91-c217-4646-8ce6-a491f0c501ea",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory 2",
            "lessonNumber": 2,
            "description": "learn second kompuktor",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T17:50:09.000Z",
            "updatedAt": "2023-02-04T17:50:09.000Z"
          },
          {
            "id": "dafd076c-cd23-42d2-aa19-c1be0fb112f0",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory 3",
            "lessonNumber": 3,
            "description": "learn kompuktor of 3 gen",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T17:50:09.000Z",
            "updatedAt": "2023-02-04T17:50:09.000Z"
          },
          {
            "id": "f1c4c340-3171-487c-ae4c-425247613457",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory 4",
            "lessonNumber": 4,
            "description": "learn kompuktor of 4 gen",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T17:50:09.000Z",
            "updatedAt": "2023-02-04T17:50:09.000Z"
          },
          {
            "id": "15304177-d4c5-4dd5-8600-de1bc3376aa3",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory 5",
            "lessonNumber": 5,
            "description": "learn kompuktor of 5 gen",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T17:50:09.000Z",
            "updatedAt": "2023-02-04T17:50:09.000Z"
          },
          {
            "id": "e434a0c7-2075-4be1-bead-ccff61dc5c17",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory",
            "lessonNumber": 6,
            "description": "learn 6 kompuktor",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T18:39:57.000Z",
            "updatedAt": "2023-02-04T18:39:57.000Z"
          },
          {
            "id": "0aa46dcb-52ff-41e1-bf13-05dc9df044b6",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory",
            "lessonNumber": 7,
            "description": "learn 6 kompuktor",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T18:40:22.000Z",
            "updatedAt": "2023-02-04T18:40:22.000Z"
          },
          {
            "id": "2709083e-85e2-41ba-aab6-c714a109ea0f",
            "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
            "title": "kompuktory",
            "lessonNumber": 8,
            "description": "learn 6 kompuktor",
            "information": [
              "information about kompuktory"
            ],
            "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
            "updatedBy": null,
            "createdAt": "2023-02-04T18:44:55.000Z",
            "updatedAt": "2023-02-04T18:44:55.000Z"
          }
        ]
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Result11"
          }
        }
      },
      "required": [
        "success",
        "result"
      ]
    },
    "Result11": {
      "title": "Result11",
      "example": {
        "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
        "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
        "title": "kompuktory",
        "lessonNumber": 1,
        "description": "learn firs kompuktor",
        "information": [
          "information about kompuktory"
        ],
        "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
        "updatedBy": null,
        "createdAt": "2023-02-04T17:50:09.000Z",
        "updatedAt": "2023-02-04T17:50:09.000Z"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "courseId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "lessonNumber": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdBy": {
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "courseId",
        "title",
        "lessonNumber",
        "description",
        "information",
        "createdBy",
        "updatedBy",
        "createdAt",
        "updatedAt"
      ]
    },
    "getlessonbyId": {
      "title": "getlessonbyId",
      "example": {
        "success": true,
        "result": {
          "id": "25fbc63c-55e3-40bb-8f5b-7bd7f8ef49a2",
          "courseId": "9e10d09e-e757-4ab5-b991-aeabd3c0d4b8",
          "title": "kompuktory",
          "lessonNumber": 1,
          "description": "learn firs kompuktor",
          "information": [
            "information about kompuktory"
          ],
          "createdBy": "a05fb7a6-ef20-4240-ac40-0e33325458b2",
          "updatedBy": null,
          "createdAt": "2023-02-04T17:50:09.000Z",
          "updatedAt": "2023-02-04T17:50:09.000Z"
        }
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "result": {
          "$ref": "#/definitions/Result11"
        }
      },
      "required": [
        "success",
        "result"
      ]
    },
    "updatelessonrequest": {
      "title": "updatelessonrequest",
      "example": {
        "title": "kompuktory",
        "description": "!!!!!!!!!!!!!!!",
        "information": [
          "information about kompuktory"
        ]
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "title",
        "description",
        "information"
      ]
    },
    "updatelesson": {
      "title": "updatelesson",
      "example": {
        "success": true
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      },
      "required": [
        "success"
      ]
    },
    "deletelesson": {
      "title": "deletelesson",
      "example": {
        "success": true
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      },
      "required": [
        "success"
      ]
    }
  },
...
}